Adjusted information view to fix of filtering views
