+++ tests/tools.vitruv.applications.viewfilter.tests/src/tools/vitruv/applications/viewfilter/tests/BasicViewFilterTest.java
@@ -122,6 +122,7 @@ public class BasicViewFilterTest extends ViewBasedVitruvApplicationTest {
 		modifyModel();
 		
 		view.update();
+		view.update();
 		
 		view.getSelection();
 	}
